requirements.txt
Django==3.1
djangorestframework==3.12
psycopg2-binary==2.8.6
Pillow
python-decouple
django-filter
drf-yasg
django-phonenumber-field==5.0.0
djangorestframework_simplejwt
twilio
django-cors-headers
django-filter
PyJWT==2.1.0